Custom Get-AWSCmdletName cmdlet.

+++ generator/02.Export-CmdletMarkdown.ps1
@@ -9,86 +9,61 @@ $cmdletVersion = Get-Module AWSPowerShell.NetCore | ForEach-Object { $_.Version.
 
 # import service metadata
 $g_SubCommandAlias = @{
-    'acm-pca'                      = 'pca';
-    'application-autoscaling'      = 'aas';
-    'application-insights'         = 'cwai';
-    'alexaforbusiness'             = 'alxb';
     'amp'                          = 'prom';
-    'apigateway'                   = 'ag';
+    'apigateway'                   = 'ag'; # API Gateway v1
     'apigatewaymanagementapi'      = 'agm';
     'apigatewayv2'                 = 'ag2';
-    'appmesh'                      = 'amsh';
     'autoscaling'                  = 'as';
-    'autoscaling-plans'            = 'asp';
-    'clouddirectory'               = 'cdir';
-    'cloudhsm'                     = 'hsm';
+    'autoscalingplans'             = 'asp';
+    'cloudhsm'                     = 'hsm';  # Cloud HSM v1
     'cloudhsmv2'                   = 'hsm2';
     'cloudsearch'                  = 'cs';
     'cloudwatch'                   = 'cw';
-    'codeguru-reviewer'            = 'cgr';
-    'codestar-notifications'       = 'cstn';
-    'cognito-identity'             = 'cgi';
-    'cognito-idp'                  = 'cgip';
-    'cognito-sync'                 = 'cgis';
+    'cognitoidentity'              = 'cgi';
+    'cognitoidp'                   = 'cgip';
+    'cognitosync'                  = 'cgis';
     'comprehend'                   = 'comp';
     'comprehendmedical'            = 'cmpm';
-    'compute-optimizer'            = 'co';
-    'configservice'                = 'cfg';
+    'computeoptimizer'             = 'co';
     'connect'                      = 'conn';
-    'datapipeline'                 = 'dp';
-    'devicefarm'                   = 'df';
-    'directconnect'                = 'dc';
     'docdb'                        = 'doc';
-    'ds-data'                      = 'dsd'; 
+    'dsdata'                       = 'dsd'; 
     'dynamodb'                     = 'ddb';
-    'elasticbeanstalk'             = 'eb';
-    'elastictranscoder'            = 'ets';
     'elbv2'                        = 'elb2';
     'events'                       = 'cwe';
-    'globalaccelerator'            = 'gacl';
-    'groundstation'                = 'gs';
-    'importexport'                 = 'ie';
-    'iot-jobs-data'                = 'iotj'
+    'iotjobsdata'                  = 'iotj'
     'iotevents'                    = 'iote';
-    'iotevents-data'               = 'ioted';
+    'ioteventsdata'                = 'ioted';
     'iotthingsgraph'               = 'iottg';
     'kinesis'                      = 'kin';
-    'kinesis-video-media'          = 'kvm';
-    'kinesis-video-signaling'      = 'kvsc';
+    'kinesisvideomedia'            = 'kvm';
+    'kinesisvideosignaling'        = 'kvsc';
     'kinesisanalytics'             = 'kina';
     'kinesisanalyticsv2'           = 'kina2';
     'kinesisvideo'                 = 'kv';
-    'lex-models'                   = 'lmb';
-    'lex-runtime'                  = 'lex';
-    'lexv2-models'                 = 'lmbv2';
-    'lexv2-runtime'                = 'lrsv2';
-    'license-manager'              = 'licm';
-    'lookoutvision'                = 'lfv';
-    'machinelearning'              = 'ml';
-    'marketplace-catalog'          = 'mcat';
-    'marketplace-entitlement'      = 'mes';
+    'lexmodels'                    = 'lmb';
+    'lexruntime'                   = 'lex';
+    'lexv2models'                  = 'lmbv2';
+    'lexv2runtime'                 = 'lrsv2';
+    'licensemanager'               = 'licm';
+    'marketplacecatalog'           = 'mcat';
+    'marketplaceentitlement'       = 'mes';
     'marketplacecommerceanalytics' = 'mca';
     'managedblockchain'            = 'mbc';
     'mediastore'                   = 'ems';
-    'mediastore-data'              = 'emsd';
+    'mediastoredata'               = 'emsd';
     'mediapackage'                 = 'emp';
-    'mediapackage-vod'             = 'empv';
+    'mediapackagevod'              = 'empv';
     'meteringmarketplace'          = 'mm';
     'mgh'                          = 'mh';
-    'migrationhub-config'          = 'mhc';
-    'opsworks'                     = 'ops';
-    'opsworks-cm'                  = 'owcm';
     'personalize'                  = 'pers';
-    'personalize-events'           = 'perse';
-    'personalize-runtime'          = 'persr';
+    'personalizeevents'            = 'perse';
+    'personalizeruntime'           = 'persr';
     'pinpoint'                     = 'pin';
-    'pinpoint-email'               = 'pine';
-    'pinpoint-sms-voice'           = 'pine'
+    'pinpointemail'                = 'pine';
+    'pinpointsmsvoice'             = 'pine'
     'pricing'                      = 'pls';
-    'privatenetworks'              = 'pv5g'
-    'qldb-session'                 = 'qldbs';
-    'rds-data'                     = 'rdsd';
-    'resource-groups'              = 'rg';
+    'resourcegroups'               = 'rg';
     'resourcegroupstaggingapi'     = 'rgt';
     'route53'                      = 'r53';
     'route53domains'               = 'r53d';
@@ -96,21 +71,12 @@ $g_SubCommandAlias = @{
     's3api'                        = 's3';
     's3control'                    = 's3c';
     'sagemaker'                    = 'sm';
-    'sagemaker-a2i-runtime'        = 'a2ir';
-    'sagemaker-runtime'            = 'smr';
-    'secretsmanager'               = 'sec';
-    'securityhub'                  = 'shub';
+    'sagemakera2iruntime'          = 'a2ir';
+    'sagemakerruntime'             = 'smr';
     'serverlessrepo'               = 'sar';
-    'service-quotas'               = 'sq';
-    'servicecatalog-appregistry'   = 'scar';
+    'servicecatalogappregistry'    = 'scar';
     'servicecatalog'               = 'sc';
-    'servicediscovery'             = 'sd';
     'sesv2'                        = 'ses2';
-    'sso-oidc'                     = 'ssooidc';
-    'stepfunctions'                = 'sfn';
-    'storagegateway'               = 'sg';
-    'waf-regional'                 = 'wafr';
-    'xray'                         = 'xr'
 }
 $g_AWSServices = Get-Content -LiteralPath "./aws-services.json" -Raw | ConvertFrom-Json -AsHashtable
 $g_CmdletReferenceLinks = Get-Content -LiteralPath "./aws-reference-links.json" -Raw | ConvertFrom-Json -AsHashtable
@@ -118,32 +84,36 @@ $g_CmdletReferenceLinks = Get-Content -LiteralPath "./aws-reference-links.json" 
 # Special operation mapping for S3API
 $g_S3APIOperationsMapping = @{
     # S3 lifecycle API is special
-    'delete-bucket-lifecycle'               = 'DeleteLifecycleConfiguration' # "DeleteBucketLifecycle" is correct, but PowerShell cmdlet returns incorrect name. 
-    'get-bucket-lifecycle'                  = '----'                         # GetBucketLifecycle is old API.
-    'get-bucket-lifecycle-configuration'    = 'GetLifecycleConfiguration'    # "GetBucketLifecycleConfiguration" is correct, but PowerShell cmdlet returns incorrect name. 
-    'put-bucket-lifecycle'                  = '----'                         # PutBucketLifecycle is old API.
-    'put-bucket-lifecycle-configuration'    = 'PutLifecycleConfiguration'    # "PutBucketLifecycleConfiguration" is correct, but PowerShell cmdlet returns incorrect name. 
+    'deletebucketlifecycle'              = 'DeleteLifecycleConfiguration' # "DeleteBucketLifecycle" is correct, but PowerShell cmdlet returns incorrect name. 
+    'getbucketlifecycle'                 = '----'                         # GetBucketLifecycle is old API.
+    'getbucketlifecycleconfiguration'    = 'GetLifecycleConfiguration'    # "GetBucketLifecycleConfiguration" is correct, but PowerShell cmdlet returns incorrect name. 
+    'putbucketlifecycle'                 = '----'                         # PutBucketLifecycle is old API.
+    'putbucketlifecycleconfiguration'    = 'PutLifecycleConfiguration'    # "PutBucketLifecycleConfiguration" is correct, but PowerShell cmdlet returns incorrect name. 
     # S3 notification API is spectial
-    'get-bucket-notification'               = '----'                  # GetBucketNotification is no longer used.
-    'get-bucket-notification-configuration' = 'GetBucketNotification' # "GetBucketNotificationConfiguration" is current, but PowerShell cmdlet returns old name.
-    'put-bucket-notification'               = '----'                  # PutBucketNotification is no longer used.
-    'put-bucket-notification-configuration' = 'PutBucketNotification' # "PutBucketNotificationConfiguration" is current, but PowerShell cmdlet returns old name.
+    'getbucketnotification'              = '----'                  # GetBucketNotification is no longer used.
+    'getbucketnotificationconfiguration' = 'GetBucketNotification' # "GetBucketNotificationConfiguration" is current, but PowerShell cmdlet returns old name.
+    'putbucketnotification'              = '----'                  # PutBucketNotification is no longer used.
+    'putbucketnotificationconfiguration' = 'PutBucketNotification' # "PutBucketNotificationConfiguration" is current, but PowerShell cmdlet returns old name.
     # PowerShell cmdlet returns incorrect operation name due to AWS SDK for .NET. 
-    'create-bucket'                         = 'PutBucket'               # "CreateBucket" is correct.
-    'create-multipart-upload'               = 'InitiateMultipartUpload' # "CreateMultipartUpload" is correct.
-    'delete-bucket-cors'                    = 'DeleteCORSConfiguration' # "DeleteBucketCors" is correct.
-    'delete-object'                         = 'DeleteObjects'           # "DeleteObject" is correct.
-    'get-bucket-cors'                       = 'GetCORSConfiguration'    # "GetBucketCors" is correct.
-    'list-object-versions'                  = 'ListVersions'            # "ListObjectVersions" is correct.
-    'put-bucket-cors'                       = 'PutCORSConfiguration'    # "PutBucketCors" is correct.
-    'upload-part-copy'                      = 'UploadPart'              # "UploadPartCopy" is correct.
-    'head-bucket'                           = 'GetObjectMetadata'       # "HeadBucket" is correct.
-    'head-object'                           = 'GetObjectMetadata'       # "HeadObject" is correct.
+    'createbucket'                       = 'PutBucket'               # "CreateBucket" is correct.
+    'createmultipartupload'              = 'InitiateMultipartUpload' # "CreateMultipartUpload" is correct.
+    'deletebucketcors'                   = 'DeleteCORSConfiguration' # "DeleteBucketCors" is correct.
+    'deleteobject'                       = 'DeleteObjects'           # "DeleteObject" is correct.
+    'getbucketcors'                      = 'GetCORSConfiguration'    # "GetBucketCors" is correct.
+    'listobjectversions'                 = 'ListVersions'            # "ListObjectVersions" is correct.
+    'putbucketcors'                      = 'PutCORSConfiguration'    # "PutBucketCors" is correct.
+    'uploadpartcopy'                     = 'UploadPart'              # "UploadPartCopy" is correct.
+    'headbucket'                         = 'GetObjectMetadata'       # "HeadBucket" is correct.
+    'headobject'                         = 'GetObjectMetadata'       # "HeadObject" is correct.
 }
 
 # clear markdown directory
-Remove-Item -Path "./markdown/*.md" -Force
-Remove-Item -Path "./markdown/post/*.md" -Force
+if ('' -eq $CliServiceName) {
+    Write-Host 'Remove ./markdown/*.md...'
+    Remove-Item -Path "./markdown/*.md" -Force
+    Write-Host 'Remove ./markdown/post/*.md...'
+    Remove-Item -Path "./markdown/post/*.md" -Force
+}
 
 # functions
 function Write-HostInfo () {
@@ -154,6 +124,58 @@ function Write-HostInfo () {
     )
     Write-Host $Message -ForegroundColor Green
 }
+
+function Get-AWSCmdletNameCustom {
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory = $true)]
+        [string]$Service,
+        [Parameter(Mandatory = $true)]
+        [string]$ApiOperation
+    )
+    # Special conversions
+    if ($null -ne $g_SubCommandAlias[$Service]) {
+        # AWS CLIのサブコマンドを直接指定してもダメなものが結構ある
+        # Some CLI subcommand dosen't return PowerShell Cmdlet information directly.
+        # So, we must to use Cmdlet prefix instead of subcommand name.($g_SubCommandAlias)
+        $Service = $g_SubCommandAlias[$Service]
+    }
+    switch ($Service) {
+        's3' {
+            if ($null -ne $g_S3APIOperationsMapping[$ApiOperation]) {
+                $ApiOperation = $g_S3APIOperationsMapping[$ApiOperation]
+            }
+            break;
+        }
+        'sso' {
+            if ($ApiOperation -eq 'login') {
+                return [PSCustomObject]@{
+                    CmdletName       = 'Invoke-AWSSSOLogin'
+                    CmdletNounPrefix = 'Invoke'
+                    ServiceName      = 'AWS Single Sign-On'
+                    ServiceOperation = 'Login'
+                }
+            }
+            if ($ApiOperation -eq 'logout') {
+                return [PSCustomObject]@{
+                    CmdletName       = 'Invoke-AWSSSOLogout'
+                    CmdletNounPrefix = 'Invoke'
+                    ServiceName      = 'AWS Single Sign-On'
+                    ServiceOperation = 'Logout'
+                }
+            }
+        }
+    }
+    # use default Get-AWSCmdletName cmdlet
+    return Get-AWSCmdletName -Service $Service -ApiOperation $ApiOperation | ForEach-Object {
+        return [PSCustomObject]@{
+            CmdletName       = $_.CmdletName
+            CmdletNounPrefix = $_.CmdletNounPrefix
+            ServiceName      = $_.ServiceName
+            ServiceOperation = $_.ServiceOperation
+        } 
+    }
+} 
 
 function Export-TopPageMarkdown {
     [CmdletBinding()]
@@ -280,9 +302,6 @@ function Get-CLISubCommands {
     )
     $commands = Get-Content -LiteralPath $CommandFilePath | ForEach-Object {
         try {
-            # AWS CLIのサブコマンドを直接指定してもダメなものが結構ある
-            # Some CLI subcommand dosen't return PowerShell Cmdlet information directly.
-            # So, we must to use Cmdlet prefix instead of subcommand name.($g_SubCommandAlias)
             $awsCLI = $_
             # split aws commandline tokens
             $cliTokens = -split $_
@@ -290,26 +309,11 @@ function Get-CLISubCommands {
                 Write-Error "Failed to parse AWS CLI command line. ($_)"
             }
             $searchService = $cliTokens[1] -replace "-", ""
-            if ($null -ne $g_SubCommandAlias[$ServiceName]) {
-                $searchService = $g_SubCommandAlias[$ServiceName]
-            }
             $searchOperation = $cliTokens[2] -replace "-", ""
-            switch ($cliTokens[1]) {
-                # Special conversions for specific services
-                's3api' {
-                    if ($null -ne $g_S3APIOperationsMapping[$cliTokens[2]]) {
-                        $searchOperation = $g_S3APIOperationsMapping[$cliTokens[2]]
-                    }
-                }
-            }
-            # Get-AWSCmdletName は1つの AwsCliCommand から複数の Cmdlet を返す場合がある
-            # どうやらAWS Tools for PowerShellのバグだった様で Ver.4.1.430 で解消された模様
-            # see : https://github.com/aws/aws-tools-for-powershell/issues/319
-            #       https://github.com/aws/aws-tools-for-powershell/commit/451df1cd87c16cf31b94ca06477945b92d8d6078
             $awsCLISubcommandName = $cliTokens[2]
             $awsCLIUrl = "https://awscli.amazonaws.com/v2/documentation/api/latest/reference/$ServiceName/$awsCLISubcommandName.html"
-            # Fix #2 : Since AWS Tools for PowerShell 4, Get-AWSCmdletName -AwsCliCommand is obsolete.
-            $cmdletNames = Get-AWSCmdletName -Service $searchService -ApiOperation $searchOperation
+            # Use custom Get-AWSCmdletName command
+            $cmdletNames = Get-AWSCmdletNameCustom -Service $searchService -ApiOperation $searchOperation
             if ($null -eq $cmdletNames) {
                 $o = [PSCustomObject]@{
                     CLI              = $awsCLI
